Remove anyhow dependency
