+++ crates/wasi-stub/src/lib.rs
@@ -74,15 +74,15 @@ pub fn stub_wasi_functions(
     binary: &[u8],
     should_stub: ShouldStub,
     return_value: u32,
-) -> anyhow::Result<Vec<u8>> {
-    let wat = wasmprinter::print_bytes(binary)?;
+) -> crate::Result<Vec<u8>> {
+    let wat = wasmprinter::print_bytes(binary).map_err(|err| std::io::Error::new(std::io::ErrorKind::Other, err))?;
     let parse_buffer = wast::parser::ParseBuffer::new(&wat)?;
 
     let mut wat: Wat = wast::parser::parse(&parse_buffer)?;
     let module = match &mut wat {
         Wat::Module(m) => m,
         Wat::Component(_) => {
-            anyhow::bail!("components are not supported")
+            return Err(Error::message("components are not supported"))
         }
     };
     let fields = match &mut module.kind {
@@ -264,3 +264,22 @@ pub fn stub_wasi_functions(
 
     Ok(module.encode()?)
 }
+
+// Error handling
+pub struct Error(Box<dyn std::error::Error + Send + Sync + 'static>);
+impl Error {
+    pub fn message(reason: impl AsRef<str>) -> Self {
+        Self(Box::new(std::io::Error::new(std::io::ErrorKind::Other, reason.as_ref().to_string())))
+    }
+}
+impl std::fmt::Debug for Error {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        std::fmt::Display::fmt(&self.0, f)
+    }
+}
+impl<E> From<E> for Error where E: std::error::Error + Send + Sync + 'static {
+    fn from(err: E) -> Self {
+        Self(Box::new(err))
+    }
+}
+pub type Result<T> = std::result::Result<T, Error>;
